requirements.txt
# Web framework and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4

# Azure Blob Storage (MAIN DEPENDENCY)
azure-storage-blob==12.23.1

# Azure Authentication (for managed identity)
azure-identity==1.19.0

# Database connectivity
pyodbc==5.1.0

# Environment and configuration
python-dotenv==1.0.1

# JSON and data handling
pydantic==2.10.3

# XML parsing (for Azure XML service and Bible XML imports)
lxml==5.3.0

# System monitoring (optional, for health checks)
psutil==6.1.0

# Testing
httpx==0.25.0

# Utilities
python-multipart==0.0.19